refactor: 예약, 좌석 상태 Enum 패키지 변경
