+++ modules/domain/src/main/java/io/github/ktg/ticketing/domain/event/model/EventSeatStatus.java
@@ -1,4 +1,4 @@
-package io.github.ktg.ticketing.domain.event;
+package io.github.ktg.ticketing.domain.event.model;
 
 public enum EventSeatStatus {
     AVAILABLE, HELD, SOLD
