+++ modules/domain/src/main/java/io/github/ktg/ticketing/domain/reservation/model/ReservationStatus.java
@@ -1,4 +1,4 @@
-package io.github.ktg.ticketing.domain.reservation;
+package io.github.ktg.ticketing.domain.reservation.model;
 
 public enum ReservationStatus {
     WAITING_PAYMENT, EXPIRED, COMPLETED, CANCELED
